[HOTSPOT-298] refactor: entities를 domains로 변경
