+++ apps/web/src/domains/notification/api/types.ts
