+++ apps/web/src/domains/notification/model/type.ts
