+++ apps/web/src/domains/policy/lib/policyDescriptionFormatter.ts
