+++ apps/web/src/domains/report/model/type.ts
