+++ apps/web/src/domains/user/ui/UserProfileIcon.tsx
