+++ apps/web/src/features/policy/policy-order/ui/OrderItem.tsx
@@ -1,8 +1,8 @@
 import { Draggable } from '@hello-pangea/dnd';
 import MoreIcon from '@hotspot/ui/assets/icons/more-2.svg';
 import { cn } from '@hotspot/ui/lib';
-import type { MemberPriority } from '@/entities/policy';
-import { UserProfileIcon } from '@/entities/user';
+import type { MemberPriority } from '@/domains/policy';
+import { UserProfileIcon } from '@/domains/user';
 import { OrderButton } from './OrderButton';
 
 interface OrderItemProps {
