+++ apps/web/src/features/policy/policy-order/ui/PolicyOrderSelector.tsx
@@ -1,4 +1,4 @@
-import { POLICY_ORDER_TYPE, type PolicyOrderType } from '@entities/policy';
+import { POLICY_ORDER_TYPE, type PolicyOrderType } from '@domains/policy';
 import More2Icon from '@hotspot/ui/assets/icons/more-2.svg';
 import UserIcon from '@hotspot/ui/assets/icons/user.svg';
 import { cn } from '@hotspot/ui/lib';
